Drop library (and hence header files, entire -devel package, etc)
The libtacacs library is poorly structured: it imports symbols from
tac_plus.c and utils.c rather than being entirely self-contained.  Some
of the imported symbols are likely to clash with commonly named symbols,
like 'debug'.

It's cleaner (for now) to statically link the library to the server
for now and not export it.  Unfortunately that means that the library
is present in both tac_plus and tac_pwd, but that's inherently more
secure anyway.